add stdio proxy
